new id to every box created

+++ ex04/js/app.js
@@ -2,6 +2,11 @@ function randomColor(){
     return Math.floor(Math.random()*256);
 }
 
+function boxToBlack(){
+    
+}
+
+ 
 function createBox(){
     const newbox = box.cloneNode()
     let r = randomColor();
@@ -10,11 +15,21 @@ function createBox(){
     let color = `rgb(${r},${g},${b})`; //`` and ${} to put variables in String
     newbox.style.backgroundColor = color;
     container.appendChild(newbox)
+    newbox.id = 'box'+count
     
     newbox.addEventListener("click",function(){
         newbox.style.backgroundColor = "black";
     });
+    
+    /*clickedBox = document.getElementById('box'+count.toString());
+    clickedBox.addEventListener("click",function(){
+            clickedBox.style.backgroundColor = "black";
+    });*/
 }
+
+    
+    
+
 
 function deleteBox(){
     container.removeChild(container.lastElementChild)
@@ -25,17 +40,20 @@ const container = document.querySelector("#container")
 const box = document.createElement("div")
 box.classList.add("box")
 
-var count = 0;
+
+var count = 0
 document.addEventListener("keydown", function(event) {
-        if (event.key === 'ArrowDown') {
+    if (event.key === 'ArrowDown') {
+        if (container.childElementCount<225){  //checking not at max number of cubes
             createBox()
             count++
-        } else if (event.key === 'ArrowUp') {
+        }
+    } else if (event.key === 'ArrowUp') {
+        if(container.childElementCount>0){      //checking at least one cube to delete
             deleteBox()
             count--
         }
-        console.log(count)
+    }   
 });
-
